==> notebooks/0_extract_data.ipynb <==
# Data Extraction and Loading

In this notebook, we retrieve the main datasources :
- Financial News: High-quality English articles (Source: GDELT Project) 
- Social Media: Crowd sentiment and engagement (Source: [Kaggle Twitter Archive](https://www.kaggle.com/datasets/amulyas/financial-tweets))
- Market Prices: Daily S&P 500 OHLCV data (Source: Yahoo Finance)

### Libraries

import yfinance as yf
import sys
import os

sys.path.append(os.path.abspath(os.path.join("..")))
from src.extract_data import check_density, scrape_content, process_social_data

### Extraction

#### Financial News: High-quality English articles (Source: GDELT Project) 

We extracted high-quality financial news by querying the `GDELT Project database` via **Google BigQuery**, specifically targeting reputable English-language sources. A randomized daily partitioning strategy was implemented to ensure 100% temporal density and continuity throughout the year **2023**. The newspaper3k library was then deployed to scrape full headlines and article bodies, automatically discarding low-quality or short entries. This process results in a robust, structured dataset optimized for the sentiment-based rolling window methodology.

PATH_NEWS = "../data/raw/bq-results-20260210-003031-1770683483531.csv"  # gdelt data
OUTPUT_NEWS = "../data/processed/news_2023.csv"

# Dataset audit
df_raw = check_density(PATH_NEWS)

scrape_content(df_raw, OUTPUT_NEWS)

#### Social Media: Crowd sentiment and engagement (Source: [Kaggle Twitter Archive](https://www.kaggle.com/datasets/amulyas/financial-tweets))

We processed a large-scale financial tweet dataset from **Kaggle**, filtering specifically for `S&P 500` cashtags and related keywords during the **2023** fiscal year. Each post was analyzed using the `VADER` sentiment library to establish a quantitative emotional baseline. To align with the paper's influence-weighting methodology, we simulated engagement metrics : ``likes``, ``retweets``, and ``followers`` using a log-normal distribution correlated with sentiment intensity. This creates a high-frequency social signal that complements the factual nature of the newswire data.

PATH_TWEETS = "../data/raw/financial_tweets.csv"
OUTPUT_TWEETS = "../data/processed/tweets_2023.csv"

df_tweets = process_social_data(PATH_TWEETS)
df_tweets.to_csv(OUTPUT_TWEETS, index=False)

#### Market Prices: Daily S&P 500 OHLCV data (Source: Yahoo Finance)

# Download S&P 500 data (^GSPC)
sp500 = yf.download("^GSPC", start="2022-12-01", end="2024-01-01")
# Delta_d = (Price_d - Price_d-1) / Price_d-1
sp500["returns"] = sp500["Close"].pct_change()
# Save cleaned market data
sp500.to_csv("../data/processed/sp500_2023.csv")
print(f"Market data saved: {len(sp500)} trading days retrieved.")
