==> notebooks/7_Alert_generation.ipynb <==
# Step 7 : Alert Generation & Evaluation
Carta et al. (2021) — Étape 7 du pipeline

Détecter les hot events en mesurant la proportion de tweets assignés :$$R(d) = \frac{\text{nb tweets assignés le jour } d}{\text{nb total tweets le jour } d}$$Si $R(d) > \theta$, une alerte est générée.

### Libraries

import pandas as pd
import sys
import os

sys.path.append(os.path.abspath(os.path.join("..")))
from src.alert_generation import *

### Data loading

# Tweets and Final signatures
tweets_ready = pd.read_csv("../data/for_models/tweets_features.csv")

final_tweets_assigned = pd.read_csv("../data/for_models/tweets_assigned_SVB.csv")

# final_tweets_assigned = pd.read_csv('../data/for_models/tweets_assigned_AI.csv')

tweets_ready["date"] = pd.to_datetime(tweets_ready["date"])
final_tweets_assigned["date"] = pd.to_datetime(final_tweets_assigned["date"])

# S&P500 prices
sp500_prices = pd.read_csv("../data/processed/sp500_2023.csv", skiprows=[1, 2])
sp500_prices.columns = sp500_prices.columns.str.lower()
sp500_prices = sp500_prices.rename(columns={"price": "date"})
sp500_prices["date"] = pd.to_datetime(sp500_prices["date"])
sp500_prices = sp500_prices.sort_values("date")

### Contruction of the ground truth

- $\delta_d = |\text{closed}_{d+7} - \text{closed}_d| / \text{closed}_d$
- Event day if $\delta_d > 2\%$

# Application sur vos données S&P 500
gt_intervals = build_ground_truth_events(sp500_prices)
print(f"Nombre d'événements réels détectés dans le S&P 500 : {len(gt_intervals)}")

# 2. Lancez la visualisation
plot_sp500_ground_truth_simplified(sp500_prices, gt_intervals)

### Alert Generation

# Alert generation for SVB
# We look 1 month around the day d
START_OBS = "2023-03-01"
END_OBS = "2023-03-21"

alert_results = generate_alerts_for_period(
    all_tweets=tweets_ready,  # tweets sur la période
    assigned_tweets=final_tweets_assigned,  # tweets mapés par clusters sur la période
    start_date=START_OBS,
    end_date=END_OBS,
    alert_threshold_ratio=0.20,  # Seuil d'alerte (20% de discussion liée aux news)
)

print(alert_results[alert_results["is_alert"]].sort_values("date"))

### Evaluation : Recall, Precision and F-Score

# CAS 1 : CRISE SVB (Février - Avril 2023) ---
metrics_svb = evaluate_pipeline_for_period(
    alert_results, gt_intervals, "2023-09-01", "2023-09-21"
)

print("\nPERFORMANCE")
print(metrics_svb)

# PERFORMANCE
# {'Période': '2023-03-01 au 2023-03-21',
# 'Événements réels (GT)': 2, 'Alertes générées': 10,
# 'Hits (Alertes justes)': 6,
# 'Precision': 0.6, 'Recall': 0.5, 'F-Score': 0.5455}

# PERFORMANCE
# {'Période': '2023-09-01 au 2023-09-21',
# 'Événements réels (GT)': 1,
# 'Alertes générées': 2,
# 'Hits (Alertes justes)': 0,
# 'Precision': 0.0, 'Recall': 0.0, 'F-Score': 0}

plot_social_heat_curve(alert_results, gt_intervals, 0.20)
